// src/api/openapi-generated/models/login-request.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Yoedu Real Estate API
 * API documentation for Yoedu Real Estate backend
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



export interface LoginRequest {
    'email': string;
    'password': string;
}